update TS version of App test and remove JS version
